Change xcodebuild to be a "slow operation" (#61518)
The Xcode build almost always takes longer than 2 seconds. With the
operation classified as a "fast operation", the Flutter tool will
warn the user that the build is "taking an unexpectedly long time".
This updates the operation to be a "slow operation", which will only
warn the user if it takes 2+ minutes.