Allow transparent barrierColor in showDialog methods (#61516)
`showDialog()` and `showGeneralDialog()` were allowing transparent
colors but then triggering an assertion down the stack while building
the modal barrier. The assertion existed to keep from animating from
transparent to transparent, but there's no need for the assertion,
since we can just treat the transparent case as the same as the null
case -- no animation necessary.